Added mock-server and fetching management
