+++ package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "axios": "^0.19.0",
     "bootstrap": "^4.3.1",
     "prop-types": "^15.7.2",
     "react": "^16.10.2",
@@ -12,11 +13,15 @@
     "recharts": "^1.7.1"
   },
   "scripts": {
+    "dev:mock-server": "mock-server --watch --delay 1000 --require @babel/register",
+    "dev:dev-server": "react-scripts start",
+    "dev": "npm-run-all -p dev:*",
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
+  "proxy": "http://localhost:3456",
   "eslintConfig": {
     "extends": "react-app"
   },
@@ -33,7 +38,9 @@
     ]
   },
   "devDependencies": {
+    "@staticdeploy/mock-server": "^1.10.0",
     "enzyme": "^3.10.0",
-    "enzyme-adapter-react-16": "^1.14.0"
+    "enzyme-adapter-react-16": "^1.14.0",
+    "npm-run-all": "^4.1.5"
   }
 }
