+++ src/App.js
@@ -1,39 +1,11 @@
-import React, {Fragment} from 'react';
+import React from 'react';
 
-import Ethereum from './components/Ethereum'
-import Bitcoin from './components/Bitcoin'
-import './App.css';
-
-const btc_data = {
-  'data1': 'data1',
-  'data2': 'data2',
-  'data3': 'data3',
-  'data4': 'data4'
-}
-const btc_graph = [
-  {"name": "Coingraph", "uv": 0, "pv": 0, "amt": 0},
-  {"name": "Coingraph", "uv": 100, "pv": 100, "amt": 100},
-  {"name": "Coingraph", "uv": 80, "pv": 80, "amt": 80}
-];
-
-const eth_data = {
-  'data1': 'data1',
-  'data2': 'data2',
-  'data3': 'data3',
-  'data4': 'data4'
-}
-const eth_graph = [
-  {"name": "Coingraph", "uv": 0, "pv": 0, "amt": 0},
-  {"name": "Coingraph", "uv": 100, "pv": 100, "amt": 100},
-  {"name": "Coingraph", "uv": 50, "pv": 50, "amt": 50}
-];
+import CoinsContainer from './containers/Coins'
+import './App.css'
 
 function App() {
   return (
-    <Fragment>
-      <Bitcoin datagraph={btc_graph} dataset={btc_data}/>
-      <Ethereum datagraph={eth_graph} dataset={eth_data}/>
-    </Fragment>
+    <CoinsContainer />
   );
 }
